Update ReadME with student created utility
I added a link to a small python utility that creates directories for all the sleuth cases (or for just pro/rookie if the user specifies).

This is my first experience with github so I am not very familiar with the norms and best practices here. Let me know if I should be doing this differently!