Improve GreetBot and other Slack slash commands
- Slash commands `resources` and `cal` create a mess in public channels,
so they now operate only as direct messages. The ability to redirect a
direct message to other users with those commands has been removed
because 1. I didn't know how to make it work right away and 2. this will
prevent potentially a lot of spam.
- GreetBot is now as succinct as possible while telling any new user
about what's really important (that's the hope at least).