Say goodbye to Deadline Manager Bot :wave:
- I believe the goal was interesting, but there is a feature in Coursera
to sync with Google Calendar that works seamlessly so far and that
should be much more accurate than whatever individual students can come
up with.
- So far in semester 2, there has been no need for a bot since
important communications are being shared in Slack in one channel or
another (mainly `#general` and `#official-communications`).
- I have no time (or so I pretend) to make sure the info shared by the
bot is accurate and it seems like nobody misses it, so I let it die
suddenly by unplugging the Zapier integration. Castigate me if I was
wrong, but I don't think so. You can revive it if you really are up to
to the challenge and are ready to bear the responsibility for its
existence and possible missteps :).