Regenerate TOCs with Vim-Markdown-TOC
- This helps with consistency since the table of contents is
automatically updated every time the file is saved when the Vim plugin
is installed. Anytime a heading may be modified, it will be picked up.
- This is especially useful when two or more items have the same
sub-heading, since those need to be numbered for the links to work
properly. This is something that also happens automatically with that
plugin, which helps to avoid broken links that won't be verified with
the current link checker script.
- Sure, other plugins work too, but I bet I'll be the main contributor
for a while so I'm just trying to make my own life a bit easier ;).