Updated a specific bit of information about Sleuth
It was mentioned in Slack that this particular bit of information about Sleuth was a little ambiguous. I have tried to clarify that, and added some more information. The inability to view image-coordinates in VS Code has been a big problem for people trying to switch to VS Code, so I have added the easiest method out of the consensus in here.
I also changed the order in which these appear.